requirements.txt
alembic==1.4.0
argh==0.26.2
bleach==2.1.3
certifi==2018.4.16
chardet==3.0.4
click==6.7
dnspython==1.16.0
et-xmlfile==1.0.1
eventlet==0.25.1
Faker==0.8.17
Flask==1.0.2
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-Moment==0.6.0
Flask-OAuthlib==0.9.5
Flask-Script==2.0.6
Flask-SocketIO==3.0.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gevent==1.4.0
gevent-websocket==0.10.1
greenlet==0.4.14
gunicorn==20.0.4
html2text==2020.1.16
html5lib==1.0.1
idna==2.7
itsdangerous==0.24
jdcal==1.4.1
Jinja2==2.10
joblib==0.14.1
Mako==1.1.1
Markdown==2.6.11
MarkupSafe==1.0
mock==2.0.0
monotonic==1.5
nltk==3.4.5
numpy==1.18.1
oauthlib==2.1.0
openpyxl==3.0.3
pandas==1.0.1
pathtools==0.1.2
pbr==4.2.0
psycopg2==2.8.4
Pygments==2.2.0
Pyphen==0.9.5
python-dateutil==2.7.3
python-dotenv==0.10.0
python-editor==1.0.4
python-engineio==3.11.2
python-socketio==4.4.0
pytz==2019.3
PyYAML==3.13
requests==2.19.1
requests-oauthlib==1.0.0
scikit-learn==0.22.1
scipy==1.4.1
six==1.11.0
sklearn==0.0
SQLAlchemy==1.2.10
text-unidecode==1.2
textstat==0.6.0
urllib3==1.23
watchdog==0.8.3
webencodings==0.5.1
Werkzeug==0.14.1
WTForms==2.2.1
xlrd==1.2.0
